requirements.txt
fastapi
uvicorn
torch
transformers
sentence-transformers
faiss-cpu
datasets
PyPDF2